Reworked tests to use ByteBuffers assembled in place in the tests.

+++ src/main/java/com/splunk/javaagent/hprof2/ControlSettingsRecord.java
@@ -8,7 +8,7 @@ public class ControlSettingsRecord extends HProfRecord {
     private boolean cpuSamplingOn;
     private short stackTraceDepth;
 
-    public ControlSettingsRecord(HProfDumpFile parent, byte typeTag, long baseTimestamp, int timestampOffset,
+    public ControlSettingsRecord(HProfDumpFile parent, byte typeTag, long baseTimestamp, long timestampOffset,
                                  boolean allocTracingOn, boolean cpuSamplingOn, short stackTraceDepth) {
         super(parent, typeTag, baseTimestamp, timestampOffset);
