+++ src/main/java/com/splunk/javaagent/hprof2/StackFrame.java
@@ -4,12 +4,12 @@ public class StackFrame {
     private String methodName;
     private String methodSignature;
     private String sourceFile;
-    private int classSerialNumber;
+    private long classSerialNumber;
     private String className;
-    private Integer lineNumber;
+    private Long lineNumber;
 
     public StackFrame(String methodName, String methodSignature, String sourceFile,
-                      int classSerialNumber, String className, Integer lineNumber) {
+                      long classSerialNumber, String className, Long lineNumber) {
         this.methodName = methodName;
         this.methodSignature = methodSignature;
         this.sourceFile = sourceFile;
@@ -21,7 +21,7 @@ public class StackFrame {
     public String getMethodName() { return methodName; }
     public String getMethodSignature() { return methodSignature; }
     public String getSourceFile() { return sourceFile; }
-    public int getClassSerialNumber() { return classSerialNumber; }
+    public long getClassSerialNumber() { return classSerialNumber; }
     public String getClassName() { return className; }
-    public Integer getLineNumber() { return lineNumber; }
+    public Long getLineNumber() { return lineNumber; }
 }
