+++ src/main/java/com/splunk/javaagent/hprof2/StackTrace.java
@@ -3,14 +3,14 @@ package com.splunk.javaagent.hprof2;
 import java.util.List;
 
 public class StackTrace {
-    private int threadSerialNumber;
+    private long threadSerialNumber;
     private List<StackFrame> frames;
 
-    public StackTrace(int threadSerialNumber, List<StackFrame> frames) {
+    public StackTrace(long threadSerialNumber, List<StackFrame> frames) {
         this.threadSerialNumber = threadSerialNumber;
         this.frames = frames;
     }
 
-    public int getThreadSerialNumber() { return this.threadSerialNumber; }
+    public long getThreadSerialNumber() { return this.threadSerialNumber; }
     public List<StackFrame> getStackFrames() { return this.frames; }
 }
